lockvault/backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
python-decouple==3.8
django-cors-headers==4.4.0
